feat: add react-router-dom, add routing schema in app
