+++ package.json
@@ -18,6 +18,7 @@
   "dependencies": {
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.18.0",
     "sass": "^1.69.5"
   },
   "devDependencies": {
